templates/requirements.txt
Flask==2.3.2
google-generativeai==0.8.3
google-ai-generativelanguage>=0.4.0
google-api-core>=2.15.0
google-auth>=2.27.0
Pillow==10.4.0
python-dotenv==1.0.1
requests==2.32.3
gunicorn==22.0.0
ipython>=8.12.2
Werkzeug>=3.0.0
Jinja2>=3.1.2
click>=8.1.7
itsdangerous>=2.1.2
MarkupSafe>=2.1.3